+++ lib/useTextRecognizer.d.ts
@@ -1,5 +1,0 @@
-import type { TextRecognizer, TextRecognizerOptions } from './specs/TextRecognizer.nitro';
-/**
- * Use a {@linkcode TextRecognizer}.
- */
-export declare function useTextRecognizer({ languages, }: TextRecognizerOptions): TextRecognizer;