+++ src/specs/TextRecognizerOutputResolution.ts
@@ -1,4 +1,0 @@
-/**
- * Controls which camera buffer resolution should be used for text scanning.
- */
-export type TextRecognizerOutputResolution = 'preview' | 'full'